+++ controller/controller.go
@@ -85,15 +85,23 @@ func (sd *SessionData) NewMessage(s *dg.Session, msg *dg.MessageCreate) {
 // CheckSyntax uses regexp from the standard library to check the message has the correct
 // prefix as defined by the `prefix` constant.
 func (c *controller) checkMessage() {
-	sd := c.sess
+	// If the message is not a command, award the user XP and ignore the rest of the function
+	isCMD := c.determineIfCmd()
+	if !isCMD {
+		c.awardXP()
+		return
+	}
 
-	err := c.checkWords()
+	// Check for banned words in the message.
+	foundBannedWord, err := c.checkWords()
 	if err != nil {
 		log.Printf("checkWords failed: %s", err)
 	}
 
-	isCMD := c.determineIfCmd()
-	if !isCMD {
+	// Stop execution if a banned word is found
+	if foundBannedWord {
+		// Remove the message
+		c.deleteMessage()
 		return
 	}
 
@@ -101,10 +109,17 @@ func (c *controller) checkMessage() {
 	c.cmdHandler()
 
 	// Remove command after the bot replies
-	err = sd.session.ChannelMessageDelete(c.msg.ChannelID, c.msg.ID)
+	c.deleteMessage()
+}
+
+func (c *controller) deleteMessage() error {
+	sd := c.sess
+
+	err := sd.session.ChannelMessageDelete(c.msg.ChannelID, c.msg.ID)
 	if err != nil {
 		log.Printf("failed to delete message after bot reply: %s", err)
 	}
+	return nil
 }
 
 // Reply takes the executed data and replies to the user. This is either in the channel